Remove useless `Error` type
We now handle most sources of errors internally. The only error we pass
to the outside is `hyper::Error`, so we can pass that  directly. Of
course, this makes it less easy to add more error types in the future,
but bumping the major version isn't a big problem right now.