Fix bug resulting in endless reloading if the proxy is slow
When the proxy does not respond within 2 second, another `setInterval`
run is triggered in the JS code, resulting in it reloading again. This
happened for example penguin was stopped and then started again,
causing the WS connection to be dropped and the JS code entering the
"try every 2s for a while" loop.