Add `--removal-debounce` for "remove" FS events, with a higher default
This should fix one pain point of mine when writing docs, running
`cargo watch -x doc` and running `penguin serve target/doc`. Cargo first
wipes the output directory clean and seem to only then start building
the docs. There is quite a bit of time between the wipe and writing all
the new files. This used to make Penguin already reload the sessions,
resulting in 404 pages. Now that "remove" events are treated as less
important, this should be fixed. I think it's generally a valid
assumption that a quick reload after only a deletion event is rarely
the intended behavior.